DRY and clang-format

+++ cpp/src/arrow/util/bit-util.cc
@@ -131,7 +131,7 @@ Status TransferBitmap(MemoryPool* pool, const uint8_t* data, int64_t offset,
     while (i + 1 > 0) {
       uint8_t cur_byte = data[byte_offset + i];
       if (invert_bits) {
-        dest[i] = ~static_cast<uint8_t>((cur_byte >> bit_offset) | carry);
+        dest[i] = static_cast<uint8_t>(~((cur_byte >> bit_offset) | carry));
       } else {
         dest[i] = static_cast<uint8_t>((cur_byte >> bit_offset) | carry);
       }
@@ -141,7 +141,7 @@ Status TransferBitmap(MemoryPool* pool, const uint8_t* data, int64_t offset,
   } else {
     if (invert_bits) {
       for (int64_t i = 0; i < num_bytes; i++) {
-        dest[i] = ~(data[byte_offset + i]);
+        dest[i] = static_cast<uint8_t>(~(data[byte_offset + i]));
       }
     } else {
       std::memcpy(dest, data + byte_offset, static_cast<size_t>(num_bytes));
@@ -230,8 +230,8 @@ void UnalignedBitmapAnd(const uint8_t* left, int64_t left_offset, const uint8_t*
 }
 
 void AlignedBitmapOr(const uint8_t* left, int64_t left_offset, const uint8_t* right,
-                      int64_t right_offset, uint8_t* out, int64_t out_offset,
-                      int64_t length) {
+                     int64_t right_offset, uint8_t* out, int64_t out_offset,
+                     int64_t length) {
   DCHECK_EQ(left_offset % 8, right_offset % 8);
   DCHECK_EQ(left_offset % 8, out_offset % 8);
 
@@ -245,8 +245,8 @@ void AlignedBitmapOr(const uint8_t* left, int64_t left_offset, const uint8_t* ri
 }
 
 void UnalignedBitmapOr(const uint8_t* left, int64_t left_offset, const uint8_t* right,
-                        int64_t right_offset, uint8_t* out, int64_t out_offset,
-                        int64_t length) {
+                       int64_t right_offset, uint8_t* out, int64_t out_offset,
+                       int64_t length) {
   auto left_reader = internal::BitmapReader(left, left_offset, length);
   auto right_reader = internal::BitmapReader(right, right_offset, length);
   auto writer = internal::BitmapWriter(out, out_offset, length);
@@ -314,14 +314,14 @@ Status BitmapAnd(MemoryPool* pool, const uint8_t* left, int64_t left_offset,
 }
 
 Status BitmapOr(MemoryPool* pool, const uint8_t* left, int64_t left_offset,
-                 const uint8_t* right, int64_t right_offset, int64_t length,
-                 int64_t out_offset, std::shared_ptr<Buffer>* out_buffer) {
+                const uint8_t* right, int64_t right_offset, int64_t length,
+                int64_t out_offset, std::shared_ptr<Buffer>* out_buffer) {
   if ((out_offset % 8 == left_offset % 8) && (out_offset % 8 == right_offset % 8)) {
     // Fast case: can use bytewise AND
     const int64_t phys_bits = length + out_offset;
     RETURN_NOT_OK(AllocateEmptyBitmap(pool, phys_bits, out_buffer));
-    AlignedBitmapOr(left, left_offset, right, right_offset,
-                    (*out_buffer)->mutable_data(), out_offset, length);
+    AlignedBitmapOr(left, left_offset, right, right_offset, (*out_buffer)->mutable_data(),
+                    out_offset, length);
   } else {
     // Unaligned
     RETURN_NOT_OK(AllocateEmptyBitmap(pool, length + out_offset, out_buffer));
@@ -348,6 +348,5 @@ Status BitmapXor(MemoryPool* pool, const uint8_t* left, int64_t left_offset,
   }
   return Status::OK();
 }
-
 
 }  // namespace arrow
